requirements.txt
flask>=2.2
requests>=2.31
apscheduler>=3.10
openai>=1.3
python-dotenv>=1.0